Hybrid+delta-DFTB+MPI and spin channel fix
Fix for sizing of Delta DFTB for hybrids with MPI (not tested for
correctness, but runs and matches serial case). It is probably not
working correctly for processor groups (which should probably be done
for real-space matrices).

Also size arrays for possible spin-orbit with hybrids in the future
(moving sizing and renaming for independent spin channels).